chore(engine): split webhooks server into event and codec files
Extract event-emission helpers (emitWebhookReceived,
emitWebhookAuthenticated, emitWebhookResponded, emitWebhookError,
emitAsyncFireDropped, publishEvent) into server_events.go.

Extract request/response codec helpers (buildRequestPayload,
decodeHandlerResponse) into server_codec.go.

server.go drops from 657 to 539 lines, well under the 800-line cap
and closer to the 400-line soft target.
