+++ engine/internal/webhooks/server.go
@@ -30,7 +30,6 @@ package webhooks
 import (
 	"context"
 	"crypto/subtle"
-	"encoding/json"
 	"fmt"
 	"io"
 	"net"
@@ -478,123 +477,6 @@ func (s *Server) resolveToken(host *extension.Host, route extension.WebhookRoute
 		}
 		return raw
 	}
-}
-
-// emitWebhookReceived publishes engine_webhook_received for the given
-// request. Engine-internal observability — consumers (desktop / iOS)
-// render an audit-log view if they want.
-func (s *Server) emitWebhookReceived(reqID string, route extension.WebhookRoute, r *http.Request) {
-	s.publishEvent(types.EngineEvent{
-		Type:           "engine_webhook_received",
-		AsyncKind:      string(asyncreg.KindWebhook),
-		AsyncID:        route.Path,
-		AsyncRequestID: reqID,
-		AsyncMethod:    r.Method,
-		AsyncPath:      route.Path,
-	})
-}
-
-func (s *Server) emitWebhookAuthenticated(reqID string, route extension.WebhookRoute, r *http.Request) {
-	s.publishEvent(types.EngineEvent{
-		Type:           "engine_webhook_authenticated",
-		AsyncKind:      string(asyncreg.KindWebhook),
-		AsyncID:        route.Path,
-		AsyncRequestID: reqID,
-		AsyncMethod:    r.Method,
-		AsyncPath:      route.Path,
-	})
-}
-
-func (s *Server) emitWebhookResponded(reqID string, route extension.WebhookRoute, r *http.Request, status int, start time.Time) {
-	s.publishEvent(types.EngineEvent{
-		Type:            "engine_webhook_responded",
-		AsyncKind:       string(asyncreg.KindWebhook),
-		AsyncID:         route.Path,
-		AsyncRequestID:  reqID,
-		AsyncMethod:     r.Method,
-		AsyncPath:       route.Path,
-		AsyncStatus:     status,
-		AsyncDurationMs: time.Since(start).Milliseconds(),
-	})
-}
-
-func (s *Server) emitWebhookError(reqID, id, method, path string, status int, reason string, start time.Time) {
-	s.publishEvent(types.EngineEvent{
-		Type:            "engine_webhook_handler_error",
-		AsyncKind:       string(asyncreg.KindWebhook),
-		AsyncID:         id,
-		AsyncRequestID:  reqID,
-		AsyncMethod:     method,
-		AsyncPath:       path,
-		AsyncStatus:     status,
-		AsyncReason:     reason,
-		AsyncDurationMs: time.Since(start).Milliseconds(),
-	})
-}
-
-func (s *Server) emitAsyncFireDropped(kind, id, reason string) {
-	s.publishEvent(types.EngineEvent{
-		Type:        "engine_async_fire_dropped",
-		AsyncKind:   kind,
-		AsyncID:     id,
-		AsyncReason: reason,
-	})
-}
-
-func (s *Server) publishEvent(ev types.EngineEvent) {
-	s.mu.RLock()
-	fn := s.emit
-	s.mu.RUnlock()
-	if fn != nil {
-		fn(ev)
-	}
-}
-
-// buildRequestPayload assembles the wire payload sent to the SDK
-// runtime for a single fire. The SDK runtime exposes lazy `req.json()`
-// / `req.text()` accessors over the rawBody bytes.
-func buildRequestPayload(r *http.Request, route extension.WebhookRoute, body []byte) map[string]interface{} {
-	headers := make(map[string]string, len(r.Header))
-	for k, v := range r.Header {
-		if len(v) > 0 {
-			// Single-valued headers are the common case; for multi-valued
-			// headers we send the first value and document it.
-			headers[k] = v[0]
-		}
-	}
-	return map[string]interface{}{
-		"method":  r.Method,
-		"path":    route.Path,
-		"url":     r.URL.String(),
-		"query":   r.URL.RawQuery,
-		"headers": headers,
-		"body":    string(body),
-		"bodyB64": false,
-		"remote":  r.RemoteAddr,
-	}
-}
-
-// decodeHandlerResponse extracts {status, body, headers} from the
-// subprocess's reply. Tolerant of missing or null fields — a bare null
-// reply (handler returned undefined) becomes 200 / "" / no headers.
-func decodeHandlerResponse(raw json.RawMessage) (int, string, map[string]string) {
-	if len(raw) == 0 || string(raw) == "null" {
-		return http.StatusOK, "", nil
-	}
-	var parsed struct {
-		Status  int               `json:"status"`
-		Body    string            `json:"body"`
-		Headers map[string]string `json:"headers"`
-	}
-	if err := json.Unmarshal(raw, &parsed); err != nil {
-		// Not a structured response — treat the whole payload as the
-		// body and return 200.
-		return http.StatusOK, string(raw), nil
-	}
-	if parsed.Status == 0 {
-		parsed.Status = http.StatusOK
-	}
-	return parsed.Status, parsed.Body, parsed.Headers
 }
 
 // authError carries an HTTP status code and human-readable message
